+++ libs/directus/directus-block/src/utils/index.ts
@@ -1,5 +1,12 @@
 import { useFragment } from '../generated/fragment-masking'
+import { getBlock } from './getBlock'
 
-export { default as getBlockProps } from './get-block-props'
-export { default as mergeConfigs } from './merge-configs'
+/**
+ * @deprecated Use `getBlock` instead
+ */
+export const useBlock = getBlock
+
+export { default as getBlockProps } from './getBlockProps'
+export { default as mergeConfigs } from './mergeConfigs'
 export { useFragment as getFragment }
+export { getBlock }