+++ libs/directus/directus-query/src/index.ts
@@ -2,3 +2,7 @@ export * from './lib/hooks'
 export * from './lib/query'
 export { logger as DirectusQueryLogger } from './logger'
 export { default as initDirectusQuery } from './lib/init'
+export {
+  graphqlRequestClient as defaultGraphqlRequestClient,
+  graphqlRequestAdmin as defaultGraphqlRequestAdmin,
+} from './lib/request'