+++ libs/directus/directus-next-component/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.7.36 (2026-01-13)
+## 1.7.37 (2026-01-14)
 
 ### 🚀 Features
 
@@ -7,6 +7,8 @@
 
 ### 🩹 Fixes
 
+- **directus-next-component:** valid directus link href error handling ([#365](https://github.com/OKAMca/stack/pull/365))
+- ⚠️  functions starting with "use" should be reserved to react hooks ([68f1f8f](https://github.com/OKAMca/stack/commit/68f1f8f))
 - **directus-next:** bump version ([e9441ad](https://github.com/OKAMca/stack/commit/e9441ad))
 - **directus-next-component:** core-lib in externals deps ([9b04ca0](https://github.com/OKAMca/stack/commit/9b04ca0))
 - update vite-plugin-dts to version 3 ([5d33c77](https://github.com/OKAMca/stack/commit/5d33c77))
@@ -17,7 +19,32 @@
 
 ### 🧱 Updated Dependencies
 
-- Updated stack-ui to 2.0.0
+- Updated stack-ui to 1.44.0
+- Updated next-component to 1.2.4
+- Updated directus-next to 1.2.11
+
+### ❤️ Thank You
+
+- Marie-Maxime Tanguay @marie-maxime
+- Pierre-Olivier Clerson @poclerson
+- poclerson
+
+## 1.7.36 (2026-01-13)
+
+### 🚀 Features
+
+- React 19 + Next.js 15 support ([#369](https://github.com/OKAMca/stack/pull/369))
+- **directus-next-component:** directus link supports search params ([21a3e46](https://github.com/OKAMca/stack/commit/21a3e46))
+
+### 🩹 Fixes
+
+- **directus-next:** bump version ([e9441ad](https://github.com/OKAMca/stack/commit/e9441ad))
+- **directus-next-component:** core-lib in externals deps ([9b04ca0](https://github.com/OKAMca/stack/commit/9b04ca0))
+- update vite-plugin-dts to version 3 ([5d33c77](https://github.com/OKAMca/stack/commit/5d33c77))
+
+### 🧱 Updated Dependencies
+
+- Updated stack-ui to 1.43.0
 - Updated next-component to 1.2.3
 
 ### ❤️ Thank You